chore: implements context to pass screen props, navigation theme and ts
