+++ package.json
@@ -58,6 +58,8 @@
     "@commitlint/config-conventional": "^17.0.2",
     "@evilmartians/lefthook": "^1.5.0",
     "@react-native/eslint-config": "^0.72.2",
+    "@react-navigation/drawer": "^6.6.6",
+    "@react-navigation/native": "^6.1.9",
     "@release-it/conventional-changelog": "^5.0.0",
     "@types/jest": "^29.5.10",
     "@types/react": "~17.0.21",
@@ -72,6 +74,10 @@
     "react": "18.2.0",
     "react-native": "0.72.6",
     "react-native-builder-bob": "^0.20.0",
+    "react-native-gesture-handler": "^2.14.0",
+    "react-native-reanimated": "^3.6.1",
+    "react-native-safe-area-context": "^4.8.2",
+    "react-native-screens": "^3.29.0",
     "release-it": "^15.0.0",
     "ts-jest": "^29.1.1",
     "typescript": "^5.0.2"
@@ -81,7 +87,13 @@
   },
   "peerDependencies": {
     "react": "*",
-    "react-native": "*"
+    "react-native": "*",
+    "@react-navigation/native": "^6.1.9",
+    "@react-navigation/drawer": "^6.6.6",
+    "react-native-gesture-handler": "^2.14.0",
+    "react-native-reanimated": "^3.6.1",
+    "react-native-safe-area-context": "^4.8.2",
+    "react-native-screens": "^3.29.0"
   },
   "workspaces": [
     "example"
