+++ src/interfaces.ts
@@ -14,23 +14,31 @@ export interface IConfig {
   props?: {[key: string]: IProps};
 }
 
-export type styles = {
+export type Styles = {
   [key: string]: string | number;
 };
 
-export interface IUIprops {
+export type WelcomeProps = {
   logo?: any;
   appName?: string;
   appDescription?: string;
   buttonText?: string;
   styles?: {
-    container?: styles;
-    logoContainer?: styles;
-    logo?: styles;
-    isoContainer?: styles;
-    isoText?: styles;
-    isoDesc?: styles;
-    button?: styles;
-    buttonText?: styles;
+    container?: Styles;
+    logoContainer?: Styles;
+    logo?: Styles;
+    isoContainer?: Styles;
+    isoText?: Styles;
+    isoDesc?: Styles;
+    button?: Styles;
+    buttonText?: Styles;
   };
+};
+
+export interface IUIprops {
+  welcome?: WelcomeProps;
 }
+
+export type RootDrawerParamList = {
+  Welcome: undefined;
+};
