+++ src/ui/context.ts
@@ -1,0 +1,4 @@
+import {createContext} from 'react';
+import {IUIprops} from '../interfaces';
+
+export const UIPropsContext = createContext<IUIprops | {}>({});
