+++ src/ui/resources.ts
@@ -1,4 +1,6 @@
 export const uiColors = {
+  white: '#FFFFFF',
+  black: '#000000',
   primary: '#4B4751',
   secondary: '#877F62',
   error: '#AC654D',
